Refactor `UseCaseExecutionScope` into context-parameters
