+++ shared/arkitekt-cr-usecases/build.gradle.kts
@@ -21,6 +21,10 @@ kotlin {
     iosArm64()
     iosSimulatorArm64()
 
+    compilerOptions {
+        freeCompilerArgs.add("-Xcontext-parameters")
+    }
+
     sourceSets {
         commonMain {
             dependencies {
