+++ shared/arkitekt-cr-usecases/src/commonTest/kotlin/app/futured/arkitekt/crusecases/base/BaseUseCaseExecutionScopeTest.kt
@@ -1,6 +1,6 @@
 package app.futured.arkitekt.crusecases.base
 
-import app.futured.arkitekt.crusecases.scope.UseCaseExecutionScope
+import app.futured.arkitekt.crusecases.scope.CoroutineScopeOwner
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.test.StandardTestDispatcher
@@ -10,7 +10,7 @@ import kotlinx.coroutines.test.setMain
 import kotlin.test.AfterTest
 import kotlin.test.BeforeTest
 
-abstract class BaseUseCaseExecutionScopeTest : UseCaseExecutionScope {
+abstract class BaseUseCaseExecutionScopeTest : CoroutineScopeOwner {
 
     private val testDispatcher = StandardTestDispatcher()
     override val viewModelScope = TestScope(testDispatcher)
