+++ shared/feature/src/commonMain/kotlin/app/futured/kmptemplate/feature/ui/firstScreen/FirstComponent.kt
@@ -1,9 +1,8 @@
 package app.futured.kmptemplate.feature.ui.firstScreen
 
+import app.futured.arkitekt.crusecases.execute
 import app.futured.factorygenerator.annotation.GenerateFactory
-import app.futured.kmptemplate.feature.domain.CounterUseCase
-import app.futured.kmptemplate.feature.domain.CounterUseCaseArgs
-import app.futured.kmptemplate.feature.domain.SyncDataUseCase
+import app.futured.kmptemplate.feature.domain.*
 import app.futured.kmptemplate.feature.ui.base.AppComponentContext
 import app.futured.kmptemplate.feature.ui.base.ScreenComponent
 import app.futured.kmptemplate.resources.MR
@@ -13,7 +12,7 @@ import com.arkivanov.essenty.lifecycle.doOnCreate
 import dev.icerock.moko.resources.format
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.update
-import kotlinx.datetime.Clock
+import kotlinx.datetime.Instant
 import org.koin.core.annotation.Factory
 import org.koin.core.annotation.InjectedParam
 import kotlin.time.Duration.Companion.milliseconds
@@ -25,6 +24,7 @@ internal class FirstComponent(
     @InjectedParam override val navigation: FirstScreenNavigation,
     private val syncDataUseCase: SyncDataUseCase,
     private val counterUseCase: CounterUseCase,
+    private val timeStampUseCase: TimeStampUseCase,
 ) : ScreenComponent<FirstViewState, FirstUiEvent, FirstScreenNavigation>(
     componentContext = componentContext,
     defaultState = FirstViewState(),
@@ -46,33 +46,49 @@ internal class FirstComponent(
         doOnCreate {
             syncData()
             observeCounter()
-            updateCreatedAtTimestamp()
+            runTimestamp()
+        }
+    }
+
+    private fun runTimestamp() {
+        timeStampUseCase.executeWithLifecycle(Unit) {
+            onNext {
+                logger.d { "Collect: $it" }
+                updateCreatedAtTimestamp(it)
+            }
+            onError {
+                logger.e { "Collect error: $it" }
+            }
         }
     }
 
     override fun onNext() = navigateToSecond()
 
-    private fun syncData() = syncDataUseCase.execute {
-        onSuccess { person ->
-            componentState.update { it.copy(randomPerson = MR.strings.first_screen_random_person.format(person.name.orEmpty())) }
-        }
-        onError { error ->
-            componentState.update { it.copy(randomPerson = MR.strings.first_screen_random_person.format("Failed to fetch")) }
-            logger.e(error) { error.message.toString() }
+    private fun syncData() {
+        syncDataUseCase.execute {
+            onSuccess { person ->
+                componentState.update { it.copy(randomPerson = MR.strings.first_screen_random_person.format(person.name.orEmpty())) }
+            }
+            onError { error ->
+                componentState.update { it.copy(randomPerson = MR.strings.first_screen_random_person.format("Failed to fetch")) }
+                logger.e(error) { error.message.toString() }
+            }
         }
     }
 
-    private fun observeCounter() = counterUseCase.execute(CounterUseCaseArgs(interval = 200.milliseconds)) {
-        onNext { count ->
-            updateCount(count)
+    private fun observeCounter() {
+        counterUseCase.execute(CounterUseCaseArgs(interval = 1000.milliseconds)) {
+            onNext { count ->
+                updateCount(count)
 
-            if (count == COUNTER_ALERT_AT_SECONDS) {
-                logger.d { "Counter reached 10" }
-                sendUiEvent(FirstUiEvent.ShowToast(MR.strings.first_screen_counter_alert.format(COUNTER_ALERT_AT_SECONDS)))
+                if (count == COUNTER_ALERT_AT_SECONDS) {
+                    logger.d { "Counter reached 10" }
+                    sendUiEvent(FirstUiEvent.ShowToast(MR.strings.first_screen_counter_alert.format(COUNTER_ALERT_AT_SECONDS)))
+                }
             }
-        }
-        onError { error ->
-            logger.e(error) { "Counter error" }
+            onError { error ->
+                logger.e(error) { "Counter error" }
+            }
         }
     }
 
@@ -80,9 +96,9 @@ internal class FirstComponent(
         componentState.update { it.copy(counter = MR.strings.first_screen_counter.format(count)) }
     }
 
-    private fun updateCreatedAtTimestamp() {
+    private fun updateCreatedAtTimestamp(now: Instant) {
         componentState.update {
-            it.copy(createdAt = MR.strings.first_screen_created_at.format(Clock.System.now().desc("Hms")))
+            it.copy(createdAt = MR.strings.first_screen_created_at.format(now.desc("Hms")))
         }
     }
 }
